Add all_types test for postgresql and get all tests passing.
