+++ src/main/java/org/blackdread/sqltojava/entity/JdlFieldEnum.java
@@ -46,7 +46,14 @@ public enum JdlFieldEnum {
     /**
      * Defined here to allow to have a string without max length set for json type of SQL that jdl does not support by default
      */
-    JSON_AS_TEXT;
+    JSON_AS_TEXT,
+
+    UUID,
+
+    /**
+     * Defined here to allow to have a string without max length set for postgres text type
+     */
+    STRING_UNBOUNDED;
 
     public String toCamelUpper() {
         return CaseFormat.UPPER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL, this.name());
