+++ src/test/resources/jdl/all_types-expected-postgresql.jdl
@@ -1,0 +1,34 @@
+entity PostgresType {
+	myBoolean Boolean,
+	myDate LocalDate,
+	myTime String minlength(8) maxlength(8) pattern(/^(([0-1]\d)|(2[0-3])):([0-5]\d):([0-5]\d)$/),
+	myTimeWithTimeZone String minlength(8) maxlength(8) pattern(/^(([0-1]\d)|(2[0-3])):([0-5]\d):([0-5]\d)$/),
+	myTimestamp Instant,
+	myTimestampWithTimeZone ZonedDateTime,
+	myReal Float,
+	myDoublePrecision Double,
+	mySmallint Integer,
+	myInteger Integer,
+	myBigint Long,
+	myMoney BigDecimal,
+	myNumeric BigDecimal,
+	myBpchar String maxlength(1),
+	myChar String maxlength(1),
+	myText String,
+	myVarchar String maxlength(255),
+	myBytea Blob,
+	myJson String,
+	myUuid UUID
+}
+
+
+
+// Relations
+
+
+// Options
+service * with serviceClass
+paginate * with pagination
+dto * with mapstruct
+filter *
+
